updated counting logic

+++ src/commands/Counting Activity/counting/initcounting.ts
@@ -13,6 +13,11 @@ module.exports = {
             message.reply(`Only Admins can run this command!`);
             return;
         }
+
+        if (!message.channel.permissionsFor(client.user.id).has("ManageWebhooks")) {
+            message.channel.send(`❌I don't Have Permissions To Send EmbedLinks!`);
+            return;
+        };
         let option = args[0].toLowerCase();
         let guildId = message.guild.id;
         let channelId = message.channel.id;
@@ -31,11 +36,17 @@ module.exports = {
 
             });
 
-
-            message.channel.send(`Enabled Counting Activity In This Server Enjoy!\nNote: NumbersOnly is Enabled! You can toggle it by doing \`${PREFIX} numbersOnly\``);
+            message.channel.setTopic(`Count to your heart's content! by OS Bot! [Admins] To disable it type "${PREFIX}" counting disable next number is 1`);
+            message.channel.send(`Enabled Counting Activity In This Server Enjoy!`);
             await newDoc.save();
             return;
         } else if(option === 'disable'){
+            let oldDoc = await countingDoc.findOne({ guildId: guildId });
+            if (!oldDoc) {
+                message.reply(`Counting Is Already Disabled In this Channel`);
+                return;
+            }
+            message.channel.setTopic("");
             message.channel.send('Disabled Counting Activity In This Server!')
             await countingDoc.deleteMany({guildId: guildId }); 
             return;
